Rename: api -> api-client
