+++ app/global-error.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { AcdApiError, LoginRequiredError, NoPermissionError, NoRequestError, NoResponseError } from '@/lib/api/errors';
+import { AcdApiError, LoginRequiredError, NoPermissionError, NoRequestError, NoResponseError } from '@/lib/api-client/errors';
 
 import { useRouter } from 'next/navigation';
 import Button from '@/component/basic/button';
