+++ app/mypage/delete-account/page.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react';
 import DeleteAccountForm, { DeleteAccountInputExtended } from './form';
 import { handleInputBuilder } from '@/lib/form/handler';
-import { apiDeleteAccount } from '@/lib/api/calls/mypage';
+import { apiDeleteAccount } from '@/lib/api-client/calls/mypage';
 import { useSessionId } from '@/lib/context/SessionIdContext';
 import { useRouter } from 'next/navigation';
