+++ lib/api-client/errors.ts
